Problem #2A Code Solution For The Robotics Take Home Test 2024
In this code solution, I use PID to determine when to stop the robot. The robot is set to go to a certain height (random value was set in this code) when the Raise Elevator Command Is called and simply goes to the lowest it can when the Lower Elevator Command is used.